+++ packages/styled-components/src/sheet/Rehydration.js
@@ -35,12 +35,7 @@ export const outputSheet = (sheet: Sheet) => {
 
     // NOTE: It's easier to collect rules and have the marker
     // after the actual rules to simplify the rehydration
-    css += rules + selector;
-    if (content.length > 0) {
-      css += `{content:"${content}"}\n`;
-    } else {
-      css += '{}\n';
-    }
+    css += `${rules}${selector}{content:"${content}"}\n`;
   }
 
   return css;
@@ -63,7 +58,10 @@ const rehydrateSheetFromTag = (sheet: Sheet, style: HTMLStyleElement) => {
   for (let i = 0, l = cssRules.length; i < l; i++) {
     const cssRule = (cssRules[i]: any);
 
-    if (cssRule.type !== PLAIN_RULE_TYPE) {
+    if (typeof cssRule.cssText !== 'string') {
+      // Avoid IE11 quirk where cssText is inaccessible on some invalid rules
+      continue;
+    } else if (cssRule.type !== PLAIN_RULE_TYPE) {
       rules.push(cssRule.cssText);
     } else {
       const marker = cssRule.selectorText.match(MARKER_RE);