+++ packages/styled-components/src/test/expanded-api.test.js
@@ -4,6 +4,10 @@ import TestRenderer from 'react-test-renderer';
 
 import { getCSS, resetStyled } from './utils';
 
+// Disable isStaticRules optimisation since we're not
+// testing for ComponentStyle specifics here
+jest.mock('../utils/isStaticRules', () => () => false);
+
 let styled;
 
 describe('expanded api', () => {